dnfdaemon: package_to_map() to separate file
